try fix sca test
